+++ test/Sentry.Testing/ModuleInit.cs
@@ -3,8 +3,6 @@ using System.Runtime.CompilerServices;
 public static class ModuleInit
 {
     [ModuleInitializer]
-    public static void Init()
-    {
+    public static void Init() =>
         VerifyDiffPlex.Initialize();
-    }
 }