+++ test/Sentry.Testing/VerifyExtensions.cs
@@ -22,6 +22,8 @@ public static class VerifyExtensions
                 _ => _.Env,
                 _ => _.Url,
                 _ => _.Headers)
+            .IgnoreMembers<SessionUpdate>(
+                _ => _.Duration)
             .IgnoreMembers<Transaction>(
                 _ => _.Release)
             .IgnoreMembers<SentryException>(